# smd-vital-backend/docs/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: SMD Vital - Microservices API
  description: |
    API completa para la plataforma de servicios médicos domiciliarios SMD Vital Bogotá.
    
    ## Servicios Disponibles
    
    - **Authentication Service**: Gestión de autenticación y autorización
    - **User Service**: Gestión de usuarios y perfiles
    - **Appointment Service**: Programación y gestión de citas médicas
    - **Notification Service**: Envío de notificaciones multicanal
    - **Medical Records Service**: Historiales médicos y documentos
    - **Payment Service**: Procesamiento de pagos y facturación
    
    ## Autenticación
    
    Todos los endpoints (excepto públicos) requieren autenticación JWT.
    Incluye el token en el header: `Authorization: Bearer <token>`
    
    ## Códigos de Estado
    
    - `200` - Éxito
    - `201` - Creado exitosamente
    - `400` - Error en la solicitud
    - `401` - No autorizado
    - `403` - Prohibido
    - `404` - No encontrado
    - `422` - Error de validación
    - `500` - Error interno del servidor
  version: 1.0.0
  contact:
    name: SMD Vital Backend Team
    email: dev@smdvitalbogota.com
    url: https://smdvitalbogota.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Desarrollo Local (API Gateway)
  - url: http://localhost:8001
    description: Auth Service Directo
  - url: http://localhost:8002
    description: User Service Directo
...
paths:
...
  /notifications:
    post:
      tags:
        - Notifications
      summary: Crear notificación
      description: Crear y enviar una nueva notificación
      operationId: createNotification
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationCreate'
      responses:
        '201':
          description: Notificación creada y enviada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
  /notifications/user/{userId}:
    get:
      tags:
        - Notifications
      summary: Obtener notificaciones del usuario
      description: Obtener notificaciones de un usuario específico
      operationId: getUserNotifications
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: ID del usuario
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: Filtrar por estado de la notificación
          schema:
            type: string
            enum: [pending, sent, read, failed]
        - name: limit
          in: query
          description: Número de notificaciones a obtener
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Lista de notificaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
  /notifications/{notificationId}/mark-read:
    put:
      tags:
        - Notifications
      summary: Marcar notificación como leída
      description: Marcar una notificación específica como leída
      operationId: markNotificationAsRead
      security:
        - BearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          description: ID de la notificación
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notificación marcada como leída
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /notifications/send-template:
    post:
      tags:
        - Notifications
      summary: Enviar notificación con plantilla
      description: Enviar notificación usando una plantilla predefinida
      operationId: sendTemplateNotification
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateNotificationRequest'
      responses:
        '201':
          description: Notificación de plantilla enviada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /preferences:
    post:
      tags:
        - Notifications
      summary: Configurar preferencias de notificación
      description: Configurar preferencias de notificación del usuario
      operationId: setUserPreferences
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPreferenceCreate'
      responses:
        '200':
          description: Preferencias actualizadas exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /preferences/{userId}:
    get:
      tags:
        - Notifications
      summary: Obtener preferencias de notificación
      description: Obtener preferencias de notificación de un usuario
      operationId: getUserPreferences
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: ID del usuario
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Preferencias de notificación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreference'
  /notifications/stats:
    get:
      tags:
        - Notifications
      summary: Estadísticas de notificaciones
      description: Obtener estadísticas de notificaciones (solo administradores)
      operationId: getNotificationStats
      security:
        - BearerAuth: []
      parameters:
        - name: start_date
          in: query
          description: Fecha de inicio
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          description: Fecha de fin
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Estadísticas de notificaciones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationStats'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtenido del endpoint de login
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          default: true
        message:
          type: string
          description: Mensaje de éxito
        data:
          type: object
          description: Datos adicionales
...
    NotificationCreate:
      type: object
      required:
        - user_id
        - type
        - title
        - message
        - channels
      properties:
        user_id:
          type: string
          format: uuid
          description: ID del usuario destinatario
        type:
          type: string
          enum: [appointment_confirmation, appointment_reminder, prescription_ready, payment_confirmation, general]
          description: Tipo de notificación
        title:
          type: string
          description: Título de la notificación
          example: "Cita Médica Confirmada"
        message:
          type: string
          description: Mensaje de la notificación
        channels:
          type: array
          items:
            type: string
            enum: [email, sms, push, whatsapp]
          description: Canales de envío
          example: ["email", "sms"]
        priority:
          type: string
          enum: [low, normal, high, urgent]
          default: normal
          description: Prioridad de la notificación
        template_data:
          type: object
          description: Datos para personalizar plantilla
        scheduled_for:
          type: string
          format: date-time
          description: Fecha programada para envío (opcional)
    Notification:
      allOf:
        - $ref: '#/components/schemas/NotificationCreate'
        - type: object
          properties:
            id:
              type: string
              format: uuid
              description: ID de la notificación
            status:
              type: string
              enum: [pending, sent, read, failed]
              description: Estado de la notificación
            created_by:
              type: string
              format: uuid
              description: ID del usuario que creó
            created_at:
              type: string
              format: date-time
              description: Fecha de creación
            sent_at:
              type: string
              format: date-time
              description: Fecha de envío
            read_at:
              type: string
              format: date-time
              description: Fecha de lectura
    TemplateNotificationRequest:
      type: object
      required:
        - template_name
        - user_id
        - template_data
        - channels
      properties:
        template_name:
          type: string
          enum: [appointment_confirmation, appointment_reminder, prescription_ready, payment_confirmation]
          description: Nombre de la plantilla
        user_id:
          type: string
          format: uuid
          description: ID del usuario destinatario
        template_data:
          type: object
          description: Datos para personalizar la plantilla
          example:
            patient_name: "Juan Pérez"
            appointment_date: "2024-01-15"
            appointment_time: "10:00 AM"
            doctor_name: "Dr. María González"
            location: "Carrera 15 # 93-01, Bogotá"
        channels:
          type: array
          items:
            type: string
            enum: [email, sms, push, whatsapp]
          description: Canales de envío
    UserPreferenceCreate:
      type: object
      required:
        - user_id
      properties:
        user_id:
          type: string
          format: uuid
          description: ID del usuario
        email_enabled:
          type: boolean
          default: true
          description: Recibir notificaciones por email
        sms_enabled:
          type: boolean
          default: true
          description: Recibir notificaciones por SMS
        push_enabled:
          type: boolean
          default: true
          description: Recibir push notifications
        whatsapp_enabled:
          type: boolean
          default: true
          description: Recibir notificaciones por WhatsApp
        notification_types:
          type: array
          items:
            type: string
            enum: [appointment_confirmation, appointment_reminder, prescription_ready, payment_confirmation, general]
          description: Tipos de notificación habilitados
    UserPreference:
      allOf:
        - $ref: '#/components/schemas/UserPreferenceCreate'
        - type: object
          properties:
            id:
              type: string
              format: uuid
              description: ID de la preferencia
            created_at:
              type: string
              format: date-time
              description: Fecha de creación
            updated_at:
              type: string
              format: date-time
              description: Fecha de actualización
    NotificationStats:
      type: object
      properties:
        period:
          type: object
          properties:
            start_date:
              type: string
              format: date-time
            end_date:
              type: string
              format: date-time
        summary:
          type: object
          properties:
            total_notifications:
              type: integer
              description: Total de notificaciones
            sent_notifications:
              type: integer
              description: Notificaciones enviadas
            failed_notifications:
              type: integer
              description: Notificaciones fallidas
            success_rate:
              type: number
              format: float
              description: Tasa de éxito (%)
        by_type:
          type: object
          additionalProperties:
            type: integer
          description: Notificaciones por tipo
...